requirements.txt
torch
torchvision
torchaudio
tqdm
pillow
opencv-python
matplotlib
gdown